updating ktx version, implementing variable moment-speed, ECS movement demo refactoring

+++ core/src/org/river/exertion/koboldQueue/time/Moment.kt
@@ -1,9 +1,9 @@
 package org.river.exertion.koboldQueue.time
 
-class Moment(val milliseconds : Long) {
+class Moment(val milliseconds : Float) {
 
     companion object {
-        val Immediate = Moment(0L)
+        val Immediate = Moment(0f)
     }
 }
